[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2024 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -9,10 +9,10 @@ typedef tuple<ll, ll, ll> tlll;
 #define xx first
 #define yy second
 
-const int r = 31, M = 1234567891;
+const ll r = 31, M = 1234567891;
 int main() {
     fast_io;
-    int n, ans = 0;
+    ll n, ans = 0;
     string s;
     cin >> n >> s;
     for (int i = 0; i < n; i++) {
